refactor: auth 관련 파일들 FSD 구조에 맞게 파일이동 (#71)
- auth 관련 파일들을 FSD 구조에 맞게 이동
- 공통컴포넌트 header를 widgets로 이동
