+++ src/app/layout.tsx
@@ -1,10 +1,10 @@
 import type { Metadata } from 'next'
 import localFont from 'next/font/local'
 import './globals.css'
-import { HeaderWrapper } from '@/features/auth'
 import Footer from '@/shared/ui/Footer'
 import { Toaster } from '@/shared/ui/Toaster'
 import Providers from '@/app/providers'
+import { HeaderWrapper } from '@/features/auth'
 
 export const metadata: Metadata = {
   title: 'StudiGo',
