+++ src/entities/auth/join/model/nickname-schema.ts
@@ -1,0 +1,24 @@
+import z from 'zod'
+import { ErrorResponseSchema } from '../../login/model/common-schema'
+
+export const NicknameCheckRequestSchema = z.object({
+  nickname: z.string().min(1),
+})
+export type NicknameCheckRequest = z.infer<typeof NicknameCheckRequestSchema>
+
+export const NicknameCheckResponseSchema = z.object({
+  available: z.boolean(),
+  message: z.string(),
+})
+export type NicknameCheckResponse = z.infer<typeof NicknameCheckResponseSchema>
+
+export const NicknameCheckErrorSchema = ErrorResponseSchema.extend({
+  error_code: z.enum([
+    'INVALID_NICKNAME_LENGTH',
+    'INVALID_NICKNAME_FORMAT',
+    'NICKNAME_ALREADY_EXISTS',
+    'TOO_MANY_REQUEST',
+  ]),
+  retry_after: z.number().optional(),
+})
+export type NicknameCheckError = z.infer<typeof NicknameCheckErrorSchema>
