+++ src/entities/auth/login/api/login.ts
@@ -1,0 +1,17 @@
+import { api } from '@/shared/api/client'
+import {
+  EmailLoginRequest,
+  EmailLoginRequestSchema,
+  EmailLoginResponse,
+  EmailLoginResponseSchema,
+} from '../model/login-schema'
+
+export const postEmailLogin = async (
+  body: EmailLoginRequest
+): Promise<EmailLoginResponse> => {
+  const payload = EmailLoginRequestSchema.parse(body)
+
+  const { data } = await api.post('/auth/login', payload)
+
+  return EmailLoginResponseSchema.parse(data)
+}
