+++ src/entities/auth/login/api/queries.ts
@@ -1,0 +1,7 @@
+import { useMutation } from '@tanstack/react-query'
+import { postKakaoOAuth } from './oauth'
+
+export const useKakaoOAuthMutation = () =>
+  useMutation({
+    mutationFn: postKakaoOAuth,
+  })
