+++ src/entities/auth/login/model/password-reset-schema.ts
@@ -1,0 +1,28 @@
+import z from 'zod'
+import { ErrorResponseSchema } from './common-schema'
+
+export const PasswordResetRequestSchema = z.object({
+  reset_token: z.string().min(1),
+  new_password: z.string().min(1),
+  new_password_confirm: z.string().min(1),
+})
+export type PasswordResetRequest = z.infer<typeof PasswordResetRequestSchema>
+
+export const PasswordResetResponseSchema = z.object({
+  message: z.string(),
+})
+export type PasswordResetResponse = z.infer<typeof PasswordResetResponseSchema>
+
+export const PasswordResetErrorSchema = ErrorResponseSchema.extend({
+  error_code: z.enum([
+    'INVALID_RESET_TOKEN',
+    'RESET_TOKEN_EXPIRED',
+    'TOKEN_ALREADY_USED',
+    'PASSWORD_MISMATCH',
+    'INVALID_PASSWORD_FORMAT',
+    'PASSWORD_SAME_AS_PREVIOUS',
+    'PASSWORD_RECENTLY_USED',
+    'TOO_MANY_REQUEST',
+  ]),
+})
+export type PasswordResetError = z.infer<typeof PasswordResetErrorSchema>
