+++ src/entities/auth/logout/api/logout.ts
@@ -1,0 +1,19 @@
+import { api } from '@/shared/api/client'
+import {
+  LogoutRequest,
+  LogoutRequestSchema,
+  LogoutResponse,
+  LogoutResponseSchema,
+} from '../model/logout-schema'
+
+export const logout = async (
+  params: Partial<LogoutRequest> = {}
+): Promise<LogoutResponse> => {
+  const body = LogoutRequestSchema.parse(params)
+
+  const response = await api.post('/auth/logout', body, {
+    withCredentials: true,
+  })
+
+  return LogoutResponseSchema.parse(response.data)
+}
