+++ src/entities/auth/logout/lib/clear-auth.ts
