--- src/features/auth/api/endpoints/oauth.ts
@@ -1,20 +1,0 @@
-import axios from 'axios'
-import {
-  KakaoOAuthRequestSchema,
-  KakaoOAuthResponseSchema,
-  type KakaoOAuthResponse,
-} from '@/features/auth/model/schema/oauth'
-
-export const postKakaoOAuth = async (input: {
-  authorization_code: string
-  redirect_uri: string
-}): Promise<KakaoOAuthResponse> => {
-  const body = KakaoOAuthRequestSchema.parse(input)
-
-  const { data } = await axios.post('/auth/oauth/kakao', body, {
-    headers: { 'Content-Type': 'application/json' },
-    withCredentials: true,
-  })
-
-  return KakaoOAuthResponseSchema.parse(data)
-}
