--- src/features/auth/api/endpoints/refresh.ts
@@ -1,19 +1,0 @@
-import { api } from '@/shared/api/client'
-import {
-  TokenRefreshRequest,
-  TokenRefreshRequestSchema,
-  TokenRefreshResponse,
-  TokenRefreshResponseSchema,
-} from '@/features/auth/api/schemas/login/refresh'
-
-export const postTokenRefresh = async (
-  body?: TokenRefreshRequest
-): Promise<TokenRefreshResponse> => {
-  const payload = TokenRefreshRequestSchema.parse(body ?? {})
-
-  const { data } = await api.post('/auth/refresh', payload, {
-    withCredentials: true,
-  })
-
-  return TokenRefreshResponseSchema.parse(data)
-}
