--- src/features/auth/api/schemas/login/index.ts
@@ -1,73 +1,0 @@
-export {
-  UserSchema,
-  TokenResponseSchema,
-  ErrorResponseSchema,
-} from '@/features/auth/api/schemas/login/common'
-
-export type {
-  User,
-  TokenResponse,
-  ErrorResponse,
-} from '@/features/auth/api/schemas/login/common'
-
-export * from '@/features/auth/api/schemas/login/login'
-
-export {
-  TokenRefreshRequestSchema,
-  TokenRefreshResponseSchema,
-  TokenRefreshErrorSchema,
-} from '@/features/auth/api/schemas/login/refresh'
-export type {
-  TokenRefreshRequest,
-  TokenRefreshResponse,
-  TokenRefreshError,
-} from '@/features/auth/api/schemas/login/refresh'
-
-export {
-  LogoutRequestSchema,
-  LogoutResponseSchema,
-  LogoutErrorSchema,
-} from '@/features/auth/api/schemas/login/logout'
-export type {
-  LogoutRequest,
-  LogoutResponse,
-  LogoutError,
-} from '@/features/auth/api/schemas/login/logout'
-
-export {
-  FindEmailSendCodeRequestSchema,
-  FindEmailSendCodeResponseSchema,
-  FindEmailVerifyRequestSchema,
-  FindEmailVerifyResponseSchema,
-} from '@/features/auth/api/schemas/login/find-email'
-export type {
-  FindEmailSendCodeRequest,
-  FindEmailSendCodeResponse,
-  FindEmailVerifyRequest,
-  FindEmailVerifyResponse,
-} from '@/features/auth/api/schemas/login/find-email'
-
-export {
-  NicknameCheckRequestSchema,
-  NicknameCheckResponseSchema,
-  NicknameCheckErrorSchema,
-} from '@/features/auth/api/schemas/login/nickname'
-export type {
-  NicknameCheckRequest,
-  NicknameCheckResponse,
-  NicknameCheckError,
-} from '@/features/auth/api/schemas/login/nickname'
-
-export {
-  PasswordResetRequestSchema,
-  PasswordResetResponseSchema,
-  PasswordResetErrorSchema,
-} from '@/features/auth/api/schemas/login/password-reset'
-export type {
-  PasswordResetRequest,
-  PasswordResetResponse,
-  PasswordResetError,
-} from '@/features/auth/api/schemas/login/password-reset'
-
-export { WithdrawalInfoResponseSchema } from '@/features/auth/api/schemas/login/withdrawal'
-export type { WithdrawalInfoResponse } from '@/features/auth/api/schemas/login/withdrawal'
