--- src/features/auth/api/schemas/login/withdrawal.ts
@@ -1,9 +1,0 @@
-import z from 'zod'
-
-export const WithdrawalInfoResponseSchema = z.object({
-  user_info: z.record(z.string(), z.unknown()),
-})
-
-export type WithdrawalInfoResponse = z.infer<
-  typeof WithdrawalInfoResponseSchema
->
