+++ src/features/auth/index.ts
@@ -1,6 +1,5 @@
-export { setupAuthInterceptors } from '@/features/auth/lib/setup-auth-interceptors'
-export { useTokenStore } from '@/features/auth/model/store/token-store'
+export { useTokenStore } from '@/entities/token/store/token-store'
 export { JoinFunnel } from '@/features/auth/join/ui'
 export { SocialCallback } from '@/features/auth/social/SocialCallback'
 export { default as LoginPage } from '@/features/auth/login/ui/page'
-export { HeaderWrapper } from '@/features/auth/ui/HeaderWrapper'
+export { HeaderWrapper } from '@/widgets/auth-header/ui/HeaderWrapper'
