--- src/features/mypage/index.ts
@@ -1,3 +1,0 @@
-import MyPage from '@/features/mypage/ui/MyPage'
-
-export { MyPage }
