--- src/shared/ui/header/index.ts
@@ -1,3 +1,0 @@
-import Header from '@/shared/ui/header/Header'
-
-export { Header }
